style: Penyesuaian style tampilan dasboard dan lainnya

+++ halaman_dasboard/data_petugas/petugas.php
@@ -16,18 +16,22 @@ $data = mysqli_fetch_all($hasil, MYSQLI_ASSOC);
 
 <aside class="sidebar">
     <div class="brand">
-    <img src="../../asett/logo_horizontal.webp" alt="logo">
+    <img src="../../asett/logo_vertikal.webp" alt="logo">
     </div>
     <ul class="nav-links">
-    <li><a href="../dasboard.php">Beranda</a></li>
-    <li><a href="../data_petugas/petugas.php">Data Petugas</a></li>
-    <li><a href="../data_peminjaman/pinjam.php">Data Peminjam</a></li>
-    <li><a href="../data_buku/buku.php">Data Buku</a></li>
+        <li><a href="../dasboard.php">Beranda</a></li>
+        <li><a href="../data_petugas/petugas.php">Data Petugas</a></li>
+        <li><a href="../data_peminjaman/pinjam.php">Data Peminjam</a></li>
+        <li><a href="../data_buku/buku.php">Data Buku</a></li>
+        <li><a href="../../halaman_Utama/index.php">Kembali</a></li>
     </ul>
 </aside>
 </div>
 <div class="main-content">
-    <h2>Data Petugas</h2>
+    <header class="topbar">
+    <h1>Data Petugas</h1>
+    <div class="user-profile">Kelola Data Petugas</div>
+    </header>
     <div class="btn-tambah">
         <a href="tambah.php">Tambah Data</a>
     </div>
@@ -49,11 +53,11 @@ $data = mysqli_fetch_all($hasil, MYSQLI_ASSOC);
             <td><?= $no + 1; ?></td>
             <td><?= $row['nama_lengkap']; ?></td>
             <td><?= $row['jabatan']; ?></td>
-            <td><?= $row['no_hp']; ?></td>
+            <td><?= 0 . $row['no_hp']; ?></td>
             <td>
             <div class="aksi">
-                    <a href="edit.php?id_petugas=<?= $row['id_petugas']; ?>">Edit</a>
-                    <a href="hapus.php?id_petugas=<?= $row['id_petugas']; ?>">Hapus</a>
+                    <a href="edit.php?id_petugas=<?= $row['id_petugas']; ?>" class="btn-edit">Edit</a>
+                    <a href="hapus.php?id_petugas=<?= $row['id_petugas']; ?>" class="btn-hapus" onclick="return confirm('Apakah Anda yakin ingin menghapus petugas <?= $row['nama_lengkap']; ?>?');">Hapus</a>
             </div>
             </td>
         </tr>
